refactor(reducers/): Adds a directory for all reducer-related components, moved existing reducer-related components into new directory, and refactored existing code to account for these changes.
